+++ classes/Controller.php
@@ -3,7 +3,7 @@
 /**
  * @package    Grav\Plugin\Login
  *
- * @copyright  Copyright (C) 2014 - 2020 RocketTheme, LLC. All rights reserved.
+ * @copyright  Copyright (C) 2014 - 2021 RocketTheme, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
 
@@ -17,9 +17,14 @@ use Grav\Common\User\Interfaces\UserCollectionInterface;
 use Grav\Common\User\Interfaces\UserInterface;
 use Grav\Common\Utils;
 use Grav\Plugin\Email\Utils as EmailUtils;
+use Grav\Plugin\Form\Form;
+use Grav\Plugin\Form\Forms;
 use Grav\Plugin\Login\Events\UserLoginEvent;
+use Grav\Plugin\Login\Invitations\Invitation;
+use Grav\Plugin\Login\Invitations\Invitations;
 use Grav\Plugin\Login\TwoFactorAuth\TwoFactorAuth;
 use Grav\Plugin\LoginPlugin;
+use RocketTheme\Toolbox\Event\Event;
 use RocketTheme\Toolbox\Session\Message;
 
 /**
@@ -536,6 +541,75 @@ class Controller
         header('Content-Type: application/json');
         echo json_encode($json_response);
         exit;
+    }
+
+    /**
+     * @return bool
+     */
+    public function taskInvite()
+    {
+        /** @var Forms $forms */
+        $forms = $this->grav['forms'] ?? null;
+        $form = $forms ? $forms->getActiveForm() : null;
+
+        /** @var Language $t */
+        $t = $this->grav['language'];
+
+        if (null === $form) {
+            $this->grav->fireEvent('onFormValidationError', new Event([
+                'form' => $form,
+                'message' => $t->translate("PLUGIN_LOGIN.INVALID_FORM"),
+            ]));
+            return false;
+        }
+
+        $data = $form->getData();
+        $emails = $data['emails'] ?? null;
+        $emails = array_unique(preg_split('/[\s,;]+/mu', $emails));
+        $emails = array_filter($emails, static function ($str) { return $str && filter_var($str, FILTER_VALIDATE_EMAIL); });
+        if (!$emails) {
+            $this->grav->fireEvent('onFormValidationError', new Event([
+                'form' => $form,
+                'message' => $t->translate("PLUGIN_LOGIN.INVALID_INVITE_EMAILS"),
+            ]));
+            return false;
+        }
+        $message = $data['message'] ?? null;
+
+        $defaults = [
+            'expiration' => 86400
+        ];
+        $invite = (array)($form->getBlueprint()->get('form/meta/invite')) + $defaults;
+
+        /** @var UserInterface $user */
+        $user = $this->grav['user'];
+        $issuer = $user->email;
+        $invitations = Invitations::getInstance();
+        $list = [];
+        foreach ($emails as $email) {
+            $data = [
+                'email' => $email,
+                'created_by' => $issuer,
+                'created_timestamp' => time(),
+                'expiration_timestamp' => time() + $invite['expiration'],
+                'account' => $invite['account']
+            ];
+
+            $invitation = new Invitation($invitations->generateToken(), $data);
+            $old = $invitations->getByEmail($email);
+            if ($old) {
+                $invitations->remove($old);
+            }
+            $invitations->add($invitation);
+            $list[] = $invitation;
+        }
+
+        $invitations->save();
+        foreach ($list as $invitation) {
+            $this->login->sendInviteEmail($invitation, $message, $user);
+        }
+
+        return true;
     }
 
     /**